+++ otter/assign/utils.py
@@ -7,6 +7,7 @@ import os
 import json
 import pprint
 import pathlib
+import shutil
 
 from glob import glob
 from contextlib import contextmanager
@@ -232,6 +233,12 @@ def run_generate_autograder(result, assignment, gs_username, gs_password, plugin
 
         if not pdf_args.get('pagebreaks', True):
             generate_args['pagebreaks'] = False
+
+    # use temp tests dir
+    if assignment.is_python and not assignment.test_files and assignment._temp_test_dir is None:
+        raise RuntimeError("Failed to create temp tests directory for Otter Generate")
+    elif assignment.is_python and not assignment.test_files:
+        generate_cmd += ["-t", str(assignment._temp_test_dir)]
     
     if assignment.is_r:
         generate_cmd += ["-l", "r"]
@@ -276,6 +283,10 @@ def run_generate_autograder(result, assignment, gs_username, gs_password, plugin
     parser = get_parser()
     args = parser.parse_args(generate_cmd)
     generate_autograder(**vars(args), assignment=assignment, plugin_collection=plugin_collection)
+
+    # clean up temp tests dir
+    if assignment._temp_test_dir is not None:
+        shutil.rmtree(str(assignment._temp_test_dir))
 
     os.chdir(curr_dir)