+++ docs/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+**Unreleased:**
+
+* Added `NotebookMetadataTestFile` for reading tests from Jupyter Notebook metadata
+* Added optional storage of tests in notebook metadata for Otter Assign
+
 **v2.2.2:**
 
 * Added `show_all_public` configuration for grading